requirements.txt
langchain==0.3.18
langchain_community==0.3.17
langchain-google-genai
python-dotenv
langchain-core==0.3.45
langgraph==0.3.21
grandalf
langchain-openai==0.2.14